refactor: split AdvancedStatsPage.tsx into Stats/ sub-components
